requirements.txt
dj-database-url==0.5.0
Django==1.11.17
django-taggit==1.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.5.3
pytz==2019.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
